Change parseInt to parseFloat for precision.

+++ js/calculator.js
@@ -1,9 +1,9 @@
 function calculator() {
   var operation = prompt("Pick a mathematical operation to perform");
   var num1 = prompt("Pick a number");
-    num1 = parseInt(num1, 10);
+    num1 = parseFloat(num1, 10);
   var num2 = prompt("Pick another number");
-    num2 = parseInt(num2, 10);
+    num2 = parseFloat(num2, 10);
   if (operation === "+") {
     alert(num1 + num2);
   } else if (operation === "-") {
